fix: gold knob acceleration overflows int8_t and flips sign on fast turns
calcNextKnobSpeed() returns between 1.0 and 3.0, and the accelerated offset is
stored in an int8_t:

    int8_t offset_accelerated = offset * encoder.calcNextKnobSpeed(offset);

Any offset of 43 or more therefore overflows and changes sign:

    offset 42 x 3.0 = 126  ->  +126
    offset 43 x 3.0 = 129  ->  -127
    offset 50 x 3.0 = 150  ->  -106
    offset 80 x 3.0 = 240  ->   -16

A slow turn produces a small offset and behaves. A fast turn produces a large
one, acceleration multiplies it past 127, and the parameter lurches the wrong
way - which is what makes fast gold-knob turns jump instead of sweep.

modEncoderAction() already takes an int32_t, so the narrowing discarded range
for no reason.

Reproduced on a stock build of main with no custom firmware: slow turns adjust
the parameter normally, fast turns jump. The January community release predates
the encoder overhaul (#4529) and does not exhibit it.

+++ src/deluge/hid/encoder_input.cpp
@@ -177,7 +177,12 @@ checkResult:
 
 				// Do it, only if
 				if (offset + modEncoderInitialTurnDirection[e] != 0) {
-					int8_t offset_accelerated = offset * encoder.calcNextKnobSpeed(offset);
+					// int32_t, not int8_t: calcNextKnobSpeed returns up to 3.0, so any offset of 43 or
+					// more overflows a signed char and FLIPS SIGN - 43*3 = 129 becomes -127. A fast
+					// turn produces exactly that: a large offset, multiplied, wrapping negative, so
+					// the parameter lurches the wrong way. modEncoderAction takes an int32_t, so the
+					// narrowing was never needed.
+					int32_t offset_accelerated = offset * encoder.calcNextKnobSpeed(offset);
 
 					getCurrentUI()->modEncoderAction(e, offset_accelerated);
